refactor: extract bundle container construction
